+++ app/src/main/java/com/wmdd/errandz/userInfoWithReviewList/UserInfoWithReviewListActivity.java
@@ -84,7 +84,7 @@ public class UserInfoWithReviewListActivity extends AppCompatActivity {
 
         if (user.getTotalRating() > 0) {
             userRatingTextView.setText(String.format("%.1f", user.getTotalRating()));
-            userRatingBar.setRating(user.getTotalRating());
+            userRatingBar.setRating((float) user.getTotalRating());
         } else {
             userRatingTextView.setText("No Rating");
             userRatingBar.setVisibility(View.GONE);
